Add API middleware and execution modules
