+++ src/autoresearch/api/__init__.py
@@ -3,22 +3,29 @@
 from __future__ import annotations
 
 from . import routing
+from .middleware import (
+    SLOWAPI_STUB,
+    RateLimitExceeded as _RateLimitExceeded,
+    dynamic_limit,
+    get_remote_address,
+    Limiter as _Limiter,
+    parse,
+)
+from .utils import (
+    RequestLogger,
+    create_request_logger,
+    get_request_logger,
+    reset_request_log,
+)
 
 create_app = routing.create_app
 app = routing.app
-SLOWAPI_STUB = routing.SLOWAPI_STUB
-RateLimitExceeded = routing.RateLimitExceeded
-dynamic_limit = routing.dynamic_limit
 config_loader = routing.config_loader
 capabilities_endpoint = routing.capabilities_endpoint
-get_remote_address = routing.get_remote_address
 limiter = routing.limiter
-parse = routing.parse
 query_endpoint = routing.query_endpoint
-create_request_logger = routing.create_request_logger
-get_request_logger = routing.get_request_logger
-RequestLogger = routing.RequestLogger
-reset_request_log = routing.reset_request_log
+RateLimitExceeded = _RateLimitExceeded
+Limiter = _Limiter
 
 __all__ = [
     "app",
@@ -30,6 +37,8 @@ __all__ = [
     "get_remote_address",
     "limiter",
     "parse",
+    "RateLimitExceeded",
+    "Limiter",
     "query_endpoint",
     "create_request_logger",
     "get_request_logger",
