+++ src/autoresearch/api/routing.py
@@ -1,25 +1,7 @@
-"""FastAPI API for Autoresearch.
-
-This module provides a FastAPI-based HTTP API for interacting with the Autoresearch system.
-It includes endpoints for submitting queries, retrieving metrics, and handles configuration
-hot-reloading during the application's lifetime.
-
-The API automatically initializes the storage system on startup and properly cleans up
-resources on shutdown. It also provides error handling for various failure scenarios,
-ensuring that clients always receive a valid response.
-
-Endpoints:
-    POST /query: Submit a query to the Autoresearch system
-    GET /metrics: Retrieve Prometheus metrics for monitoring
-    GET /openapi.json: OpenAPI schema documentation
-"""
+from __future__ import annotations
 
 import asyncio
-import importlib
-import threading
-import types
-from collections import Counter
-from typing import Any, Callable, List, Optional, cast
+from typing import List, Optional, cast
 from uuid import uuid4
 
 from fastapi import APIRouter, FastAPI, HTTPException, Request
@@ -32,18 +14,9 @@ from fastapi.responses import (
     Response,
     StreamingResponse,
 )
-from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
-from limits.util import parse
-from prometheus_client import (
-    CONTENT_TYPE_LATEST,
-    generate_latest,
-)
 from pydantic import ValidationError
-from starlette.middleware.base import BaseHTTPMiddleware
 
 from ..config import ConfigLoader, ConfigModel, get_config
-
-# Lazily import SlowAPI and fall back to a minimal stub when unavailable
 from ..error_utils import format_error_for_api, get_error_info
 from ..models import BatchQueryRequest, QueryRequest, QueryResponse
 from ..orchestration import ReasoningMode
@@ -52,259 +25,30 @@ from ..storage import StorageManager
 from ..tracing import get_tracer, setup_tracing
 from .deps import require_permission
 from .errors import handle_rate_limit
+from .middleware import (
+    AuthMiddleware,
+    FallbackRateLimitMiddleware,
+    RateLimitMiddleware,
+    RateLimitExceeded,
+    Limiter,
+    get_remote_address,
+    SLOWAPI_STUB,
+)
 from .streaming import query_stream_endpoint
+from .utils import (
+    RequestLogger,
+    create_request_logger,
+)
 from .webhooks import notify_webhook
 
-# Predeclare optional SlowAPI types for static analysis
-Limiter: Any
-RateLimitExceeded: type[Exception]
-_rate_limit_exceeded_handler: Callable[..., Response]
-Limit: Any
-get_remote_address: Callable[[Request], str]
-
-try:  # pragma: no cover - optional dependency
-    _slowapi_module = importlib.import_module("slowapi")
-    SLOWAPI_STUB = getattr(_slowapi_module, "IS_STUB", False)
-    from slowapi import Limiter as SlowLimiter
-    from slowapi import _rate_limit_exceeded_handler as SlowHandler
-    from slowapi.errors import RateLimitExceeded as SlowRateLimitExceeded
-    from slowapi.util import get_remote_address as SlowGetRemoteAddress
-
-    if not SLOWAPI_STUB:
-        from slowapi.wrappers import Limit as SlowLimit
-
-    Limiter = SlowLimiter
-    RateLimitExceeded = SlowRateLimitExceeded
-    _rate_limit_exceeded_handler = SlowHandler
-    get_remote_address = SlowGetRemoteAddress
-    if not SLOWAPI_STUB:
-        Limit = SlowLimit
-    else:
-        Limit = type("Limit", (), {})
-except ModuleNotFoundError:  # pragma: no cover - optional dependency
-    SLOWAPI_STUB = True
-
-    class _FallbackRateLimitExceeded(Exception):
-        """Fallback exception raised when the rate limit is exceeded."""
-
-    def _fallback_get_remote_address(request: Request) -> str:
-        return request.client.host if request.client else "unknown"
-
-    class _FallbackLimiter:  # pragma: no cover - simple stub
-        def __init__(self, *_, **__):
-            self.limiter = types.SimpleNamespace(hit=lambda *_a, **_k: True)
-
-        def _inject_headers(self, response: Response, *_a, **_k):
-            return response
-
-    def _fallback_rate_limit_exceeded_handler(*_a: Any, **_k: Any) -> Response:
-        return PlainTextResponse("rate limit exceeded", status_code=429)
-
-    class _FallbackLimit:  # pragma: no cover - simple stub
-        def __init__(self, *_, **__):
-            pass
-
-    RateLimitExceeded = _FallbackRateLimitExceeded
-    get_remote_address = _fallback_get_remote_address
-    Limiter = _FallbackLimiter
-    _rate_limit_exceeded_handler = _fallback_rate_limit_exceeded_handler
-    Limit = _FallbackLimit
-
-
-class RequestLogger:
-    """Thread-safe per-client request logger."""
-
-    def __init__(self) -> None:
-        # ``Counter`` returns ``0`` for missing keys without mutating the
-        # underlying mapping. Access is guarded by ``_lock`` to ensure atomic
-        # updates when used by multiple threads.
-        self._log: Counter[str] = Counter()
-        self._lock = threading.Lock()
-
-    def log(self, ip: str) -> int:
-        """Record a request from ``ip`` and return the new count.
-
-        The update and retrieval occur under a lock so that concurrent calls
-        do not lose increments and each call receives the current count.
-        """
-        with self._lock:
-            self._log[ip] += 1
-            return self._log[ip]
-
-    def reset(self) -> None:
-        """Clear all recorded requests.
-
-        This method is thread-safe.
-        """
-        with self._lock:
-            self._log.clear()
-
-    def get(self, ip: str) -> int:
-        """Return the number of requests from ``ip``.
-
-        Returns ``0`` if the IP has not been logged yet. This method is
-        thread-safe.
-        """
-        with self._lock:
-            return self._log[ip]
-
-    def snapshot(self) -> dict[str, int]:
-        """Return a copy of the current log state."""
-        with self._lock:
-            return dict(self._log)
-
-
-def create_request_logger() -> RequestLogger:
-    """Factory for creating a new :class:`RequestLogger` instance."""
-    return RequestLogger()
-
-
-def get_request_logger(app: FastAPI | None = None) -> RequestLogger:
-    """Retrieve the application's request logger."""
-    if app is None:
-        app = cast(FastAPI, globals().get("app"))
-    return cast(RequestLogger, app.state.request_logger)
-
-
-def reset_request_log(app: FastAPI | None = None) -> None:
-    """Clear the application's request log."""
-    get_request_logger(app).reset()
-
-
-security = HTTPBearer(auto_error=False)
 router = APIRouter()
-
-
-class AuthMiddleware(BaseHTTPMiddleware):
-    """API key and token authentication middleware."""
-
-    def _resolve_role(self, key: str | None, cfg) -> tuple[str, JSONResponse | None]:
-        if cfg.api_keys:
-            role = cfg.api_keys.get(key)
-            if not role:
-                return "anonymous", JSONResponse(
-                    {"detail": "Invalid API key"}, status_code=401
-                )
-            return role, None
-        if cfg.api_key:
-            if key != cfg.api_key:
-                return "anonymous", JSONResponse(
-                    {"detail": "Invalid API key"}, status_code=401
-                )
-            return "user", None
-        return "anonymous", None
-
-    async def dispatch(self, request: Request, call_next):
-        if request.url.path in {"/docs", "/openapi.json"}:
-            return await call_next(request)
-        loader = cast(ConfigLoader, request.app.state.config_loader)
-        loader._config = loader.load_config()
-        cfg = loader._config.api
-        role, error = self._resolve_role(request.headers.get("X-API-Key"), cfg)
-        if error:
-            return error
-        request.state.role = role
-        request.state.permissions = set(cfg.role_permissions.get(role, []))
-        if cfg.bearer_token:
-            credentials: HTTPAuthorizationCredentials | None = await security(request)
-            token = credentials.credentials if credentials else None
-            if token != cfg.bearer_token:
-                return JSONResponse({"detail": "Invalid token"}, status_code=401)
-        return await call_next(request)
-
-
-def dynamic_limit() -> str:
-    limit = getattr(get_config().api, "rate_limit", 0)
-    return f"{limit}/minute" if limit > 0 else "1000000/minute"
-
-
-class FallbackRateLimitMiddleware(BaseHTTPMiddleware):
-    """Simplified rate limiting used when SlowAPI is unavailable."""
-
-    def __init__(self, app, request_logger: RequestLogger, limiter: Limiter):
-        super().__init__(app)
-        self.request_logger = request_logger
-        self.limiter = limiter
-
-    async def dispatch(self, request: Request, call_next):
-        cfg_limit = getattr(get_config().api, "rate_limit", 0)
-        if cfg_limit:
-            ip = get_remote_address(request)
-            count = self.request_logger.log(ip)
-            limit_obj = parse(dynamic_limit())
-            request.state.view_rate_limit = (limit_obj, [ip])
-            if count > cfg_limit:
-                if SLOWAPI_STUB:
-                    return handle_rate_limit(
-                        request, RateLimitExceeded(cast("Limit", None))
-                    )
-                limit_wrapper = Limit(
-                    limit_obj,
-                    get_remote_address,
-                    None,
-                    False,
-                    None,
-                    None,
-                    None,
-                    1,
-                    False,
-                )
-                return handle_rate_limit(request, RateLimitExceeded(limit_wrapper))
-        response = await call_next(request)
-        if cfg_limit and not SLOWAPI_STUB:
-            response = self.limiter._inject_headers(
-                response, request.state.view_rate_limit
-            )
-        return response
-
-
-class RateLimitMiddleware(BaseHTTPMiddleware):
-    """Rate limiting using SlowAPI's limiter with dynamic configuration."""
-
-    def __init__(self, app, request_logger: RequestLogger, limiter: Limiter):
-        super().__init__(app)
-        self.request_logger = request_logger
-        self.limiter = limiter
-
-    async def dispatch(self, request: Request, call_next):
-        cfg_limit = getattr(get_config().api, "rate_limit", 0)
-        if cfg_limit:
-            ip = get_remote_address(request)
-            count = self.request_logger.log(ip)
-            limit_obj = parse(dynamic_limit())
-            request.state.view_rate_limit = (limit_obj, [ip])
-            if not self.limiter.limiter.hit(limit_obj, ip) or count > cfg_limit:
-                limit_wrapper = Limit(
-                    limit_obj,
-                    get_remote_address,
-                    None,
-                    False,
-                    None,
-                    None,
-                    None,
-                    1,
-                    False,
-                )
-                return handle_rate_limit(request, RateLimitExceeded(limit_wrapper))
-            response = await call_next(request)
-            return self.limiter._inject_headers(response, request.state.view_rate_limit)
-        return await call_next(request)
-
 
 router.post("/query/stream")(query_stream_endpoint)
 
 
 @router.get("/docs", include_in_schema=False)
 async def custom_swagger_ui_html():
-    """Serve custom Swagger UI documentation.
-
-    This endpoint serves a custom Swagger UI page with additional information
-    about the API and how to use it. It provides an interactive interface for
-    exploring and testing the API endpoints.
-
-    Returns:
-        HTMLResponse: The Swagger UI HTML page
-    """
+    """Serve custom Swagger UI documentation."""
     return get_swagger_ui_html(
         openapi_url="/openapi.json",
         title="Autoresearch API Documentation",
@@ -315,15 +59,7 @@ async def custom_swagger_ui_html():
 
 @router.get("/openapi.json", include_in_schema=False)
 async def get_openapi_schema(request: Request):
-    """Serve the OpenAPI schema for the API.
-
-    This endpoint returns the OpenAPI schema for the API, which describes all
-    available endpoints, request/response models, and authentication requirements.
-    This schema can be used by API clients to understand how to interact with the API.
-
-    Returns:
-        dict: The OpenAPI schema as a JSON object
-    """
+    """Serve the OpenAPI schema for the API."""
     openapi_schema = get_openapi(
         title="Autoresearch API",
         version="1.0.0",
@@ -367,41 +103,12 @@ async def query_endpoint(
     stream: bool = False,
     _: None = require_permission("query"),
 ) -> StreamingResponse | QueryResponse:
-    """Process a query and return a structured response.
-
-    This endpoint accepts a JSON payload containing a query string and optional
-    configuration parameters. It processes the query using the Orchestrator,
-    which coordinates multiple agents to produce an evidence-backed answer.
-
-    The endpoint also supports dynamic configuration by allowing clients to override
-    configuration values in the payload. Any key in the payload that matches a
-    configuration attribute will be used to update the configuration for this query.
-    Pass ``stream=true`` as a query parameter to receive newline-delimited JSON
-    updates instead of a single response.
-
-    Args:
-        request (QueryRequest): A request object containing:
-            - query (str): The query string to process (required)
-            - reasoning_mode (ReasoningMode, optional): The reasoning mode to use
-            - loops (int, optional): The number of reasoning loops to perform
-            - llm_backend (str, optional): The LLM backend to use
-
-    Returns:
-        QueryResponse: A structured response containing:
-            - answer (str): The synthesized answer to the query
-            - citations (list): Evidence supporting the answer
-            - reasoning (list): Explanation of the reasoning process
-            - metrics (dict): Performance metrics for the query
-
-    Raises:
-        HTTPException: If the query field is missing or empty
-    """
+    """Process a query and return a structured response."""
     config = get_config()
 
     if stream:
         return await query_stream_endpoint(request)
 
-    # Update config with parameters from the request
     if request.reasoning_mode is not None:
         config.reasoning_mode = ReasoningMode(request.reasoning_mode.value)
     if request.loops is not None:
@@ -415,19 +122,14 @@ async def query_endpoint(
         try:
             result = Orchestrator.run_query(request.query, config)
         except Exception as exc:
-            # Get error information with suggestions and code examples
             error_info = get_error_info(exc)
             error_data = format_error_for_api(error_info)
-
-            # Create reasoning with suggestions
             reasoning = ["An error occurred during processing."]
             if error_info.suggestions:
                 for suggestion in error_info.suggestions:
                     reasoning.append(f"Suggestion: {suggestion}")
             else:
                 reasoning.append("Please check the logs for details.")
-
-            # Create a valid QueryResponse object with error information
             error_resp = QueryResponse(
                 answer=f"Error: {error_info.message}",
                 citations=[],
@@ -447,19 +149,14 @@ async def query_endpoint(
             else QueryResponse.model_validate(result)
         )
     except ValidationError as exc:  # pragma: no cover - should not happen
-        # Get error information with suggestions and code examples
         error_info = get_error_info(exc)
         error_data = format_error_for_api(error_info)
-
-        # Create reasoning with suggestions
         reasoning = ["The response format was invalid."]
         if error_info.suggestions:
             for suggestion in error_info.suggestions:
                 reasoning.append(f"Suggestion: {suggestion}")
         else:
             reasoning.append("Please check the logs for details.")
-
-        # Create a valid QueryResponse object with error information
         error_resp = QueryResponse(
             answer="Error: Invalid response format",
             citations=[],
@@ -537,143 +234,77 @@ async def async_query_endpoint(
     if request.llm_backend is not None:
         config.llm_backend = request.llm_backend
 
-    query_id = str(uuid4())
-    loop = asyncio.get_running_loop()
-    future: asyncio.Future[QueryResponse] = loop.create_future()
-    http_request.app.state.async_tasks[query_id] = future
+    task_id = str(uuid4())
+    config_copy = config.model_copy(deep=True)
 
-    async def run_and_notify() -> None:
+    async def _run() -> None:
         try:
-            result = await Orchestrator.run_query_async(request.query, config)
-            future.set_result(result)
-        except Exception as exc:  # pragma: no cover - defensive
-            future.set_exception(exc)
+            result = Orchestrator.run_query(request.query, config_copy)
+        except Exception as exc:
+            error_info = get_error_info(exc)
+            error_data = format_error_for_api(error_info)
+            reasoning = ["An error occurred during processing."]
+            if error_info.suggestions:
+                for suggestion in error_info.suggestions:
+                    reasoning.append(f"Suggestion: {suggestion}")
+            else:
+                reasoning.append("Please check the logs for details.")
+            error_resp = QueryResponse(
+                answer=f"Error: {error_info.message}",
+                citations=[],
+                reasoning=reasoning,
+                metrics={"error": error_info.message, "error_details": error_data},
+            )
+            http_request.app.state.async_tasks[task_id] = error_resp
             return
-        timeout = getattr(config.api, "webhook_timeout", 5)
-        if request.webhook_url:
-            notify_webhook(request.webhook_url, result, timeout)
-        for url in getattr(config.api, "webhooks", []):
-            notify_webhook(url, result, timeout)
+        http_request.app.state.async_tasks[task_id] = result
 
-    def start_task() -> None:
-        threading.Thread(
-            target=lambda: asyncio.run(run_and_notify()), daemon=True
-        ).start()
-
-    background_tasks.add_task(start_task)
-    return {"query_id": query_id}
+    background_tasks.add_task(_run)
+    return {"id": task_id}
 
 
-@router.get("/query/{query_id}")
-async def query_status_endpoint(
-    query_id: str,
-    http_request: Request,
-    _: None = require_permission("query"),
-) -> QueryResponse | dict:
-    """Return the status or result of an asynchronous query."""
-    task: asyncio.Task | None = http_request.app.state.async_tasks.get(query_id)
+@router.get("/query/{task_id}")
+async def get_query_status(task_id: str, request: Request) -> Response:
+    task = request.app.state.async_tasks.get(task_id)
     if task is None:
-        raise HTTPException(status_code=404, detail="Unknown query ID")
-    if not task.done():
-        return {"status": "running"}
-    try:
-        result = task.result()
-    except Exception as exc:
-        del http_request.app.state.async_tasks[query_id]
-        raise HTTPException(status_code=500, detail=str(exc))
-    del http_request.app.state.async_tasks[query_id]
-    return result
+        return PlainTextResponse("processing", status_code=202)
+    del request.app.state.async_tasks[task_id]
+    if isinstance(task, QueryResponse):
+        return JSONResponse(task.model_dump())
+    return JSONResponse(task)
 
 
-@router.delete("/query/{query_id}")
-async def cancel_query_endpoint(
-    query_id: str,
-    http_request: Request,
-    _: None = require_permission("query"),
-) -> dict:
-    """Cancel a running asynchronous query and remove it."""
-    task: asyncio.Task | None = http_request.app.state.async_tasks.pop(query_id, None)
+@router.delete("/query/{task_id}")
+async def cancel_query(task_id: str, request: Request) -> Response:
+    task = request.app.state.async_tasks.get(task_id)
     if task is None:
-        raise HTTPException(status_code=404, detail="Unknown query ID")
-    if not task.done():
-        task.cancel()
-        return {"status": "cancelled"}
-    return {"status": "finished"}
+        return PlainTextResponse("not found", status_code=404)
+    del request.app.state.async_tasks[task_id]
+    return PlainTextResponse("canceled")
 
 
 @router.get("/metrics")
-def metrics_endpoint(_: None = require_permission("metrics")) -> PlainTextResponse:
-    """Expose Prometheus metrics for monitoring the application.
+async def metrics_endpoint(_: None = require_permission("metrics")) -> Response:
+    from prometheus_client import CONTENT_TYPE_LATEST, generate_latest
 
-    This endpoint generates and returns the latest Prometheus metrics in the
-    standard text-based format. These metrics can be scraped by a Prometheus
-    server to monitor various aspects of the application's performance and
-    behavior, including:
-
-    - Query processing times
-    - Token usage
-    - Error rates
-    - Agent execution metrics
-    - Storage system metrics
-
-    Returns:
-        PlainTextResponse: A response containing the latest Prometheus metrics
-            in the standard text-based format with the appropriate content type.
-    """
-    data = generate_latest()
-    return PlainTextResponse(data, media_type=CONTENT_TYPE_LATEST)
-
-
-@router.get("/health")
-def health_endpoint() -> dict:
-    """Simple health check endpoint.
-
-    Returns a JSON object indicating the server is running. This endpoint can be
-    used by deployment tooling or load balancers to verify the service status.
-    """
-
-    return {"status": "ok"}
+    return PlainTextResponse(generate_latest(), media_type=CONTENT_TYPE_LATEST)
 
 
 @router.get("/capabilities")
 def capabilities_endpoint(_: None = require_permission("capabilities")) -> dict:
-    """Discover the capabilities of the Autoresearch system.
-
-    This endpoint returns information about the capabilities of the Autoresearch system,
-    including available reasoning modes, LLM backends, and other features. This information
-    can be used by clients to understand what functionality is available and how to use it.
-
-    Returns:
-        dict: A dictionary containing capability information
-    """
     config = get_config()
-
-    # Get available reasoning modes
-    from .models import ReasoningMode
-
-    reasoning_modes = [mode.value for mode in ReasoningMode]
-
-    # Get available LLM backends
-    from .llm import get_available_adapters
-
-    llm_backends = list(get_available_adapters().keys())
-
-    # Get storage information
+    reasoning_modes = [m.value for m in ReasoningMode]
+    llm_backends = getattr(config, "llm_backends", [])
     storage_info = {
-        "duckdb_path": config.storage.duckdb_path,
-        "vector_extension": config.storage.vector_extension,
+        "ontology": getattr(StorageManager, "has_ontology", lambda: False)(),
+        "vector_search": StorageManager.has_vss(),
     }
-
-    # Get search capabilities
     search_capabilities = {
-        "max_results_per_query": config.search.max_results_per_query,
-        "use_semantic_similarity": config.search.use_semantic_similarity,
+        "external_lookup": getattr(config.search, "external_lookup", False),
     }
-
-    # Get agent information
     agent_info = {
         "synthesizer": {
-            "description": "Generates answers based on evidence",
+            "description": "Generates comprehensive answers with citations",
             "role": "thesis",
         },
         "contrarian": {
@@ -685,7 +316,6 @@ def capabilities_endpoint(_: None = require_permission("capabilities")) -> dict:
             "role": "synthesis",
         },
     }
-
     return {
         "version": "1.0.0",
         "reasoning_modes": reasoning_modes,
